Add repositories kinds filter to tracker
In addition to ask the tracker to only process some repositories by providing their name, it's now also possible to only process the repositories of the kinds provided.

This is handy when, for any reason, it's necessary to reindex all packages from repositories of a given kind (usually used in combination with the `bypassDigest` flag).

Closes #573

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>