trivy: clean environment variables when launching scanner
In the docker.io based case, the code was trying to add the environment
variables to include the dockerhub access credentials to avoid
rate-limiting. This however leads to an immediate failure with an error
message like "$HOME is not defined" instead of being able to scan
the image.

In all other cases we were passing down the full environment, which is
also not perfect. So reset environment to the known safe variables and
append the dockerhub access credentials when needed.

Signed-off-by: Dirk Mueller <dirk@dmllr.de>