Allow setting trivy and helm cache dir
Some environments do not have $HOME set or do not grant write
permissions for that directory. This commit allows setting
TRIVY_CACHE_DIR (for the scanner) and XDG_CACHE_HOME (for hub, scanner,
and tracker) to avoid the default cache directory `$HOME/.cache`.
The variables can be set by setting the value `cacheDir` in the
`values.yaml` for the repsective component.
The environment variables will then be set and the given directory path
will be mounted as ephemeral volume (emptyDir).

Signed-off-by: Alexander Rashed <alexander.rashed@gmail.com>