Fix default image tag
The docker images are prefixed with a "v". If the imageTag value is not
explicitly set, it defaults to the appVersion.
The appVersion is a semantic version, the docker image tags are prefixed
with a "v" (f.e. "v0.14.1). This commit ensures that the default tag
uses the docker tag prefix.

Signed-off-by: Alexander Rashed <alexander.rashed@gmail.com>