Improve scanner errors logging
Errors returned from `trivy` are now added to the scanning errors logs
as-is. In addition to this, no security report for a packaged will be
generated if the scanning of one of the images fails for any reason.
Before, when the error was `image not found`, we were still generating
the report, which could lead to incomplete and misleading reports.

Closes #1387

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>