Bump Trivy to 0.19.1 + some refactoring in scanner (#1437)
Trivy is [migrating to a new JSON schema](https://github.com/aquasecurity/trivy/discussions/1050). Artifact Hub stores the reports generated by Trivy in JSON format in the database, so we are migrating to the new format as well. Even though new security reports will use the new format, there might be reports stored in the database still using the previous format that may never be updated. The UI will know how to handle both, but users using the security report API endpoint need to be aware that they might receive any of them.

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Signed-off-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>
Co-authored-by: Sergio Castaño Arteaga <tegioz@icloud.com>
Co-authored-by: Cintia Sanchez Garcia <cynthiasg@icloud.com>