Auto-focus search form input field
This sets the autoFocus attribute on the input field of the search form.
It does it only when the search form is displayed on the home page, not
on the search result page.

It also sets the autoFocus attribute on the Email field of the LogIn
Modal, for two reasons:

- convenience (field gets automatically focused when trying to log in)
- perhaps security (we're worried that the login modal might pop up
  spontaneously, for instance if the user's session expires, and the
  user might start typing their credentials while the search form is
  still selected)

Note: I haven't been able to test that latter behavior (I don't know
how to trigger the login modal to appear without also interacting
with the mouse, which would change the focus anyway...) so I don't
know if this is strictly necessary, but the convenience improvement
might be useful anyway.

Signed-off-by: Jérôme Petazzoni <jerome.petazzoni@gmail.com>