Introduce package alternative name concept (#2696)
Sometimes a package can be identified by two similar names. Some
examples would be *postgres* / *postgresql* or *mongodb* / *mongo*.
Users often may type any of the options and expect the same results.
When searching for packages, Artifact Hub gives preference to **exact**
matches in names, so sometimes the top results may not be what users
would expect. This situation can be improved by providing an alternative
name for your package, which will be given the same weight as the
package name when indexing. So in cases like the previous examples, it
can help ranking them higher in the search results.

Related to #2632

Signed-off-by: Sergio Castaño Arteaga <tegioz@icloud.com>